Remove old Decodable support
