+++ Sources/Core/IncitoDocument.swift
@@ -123,8 +123,8 @@ public struct LayoutProperties {
 
 public struct TextViewProperties {
     
-    struct Span: Decodable {
-        enum SpanType: String, Decodable {
+    struct Span {
+        enum SpanType: String {
             case superscript
         }
         
@@ -133,7 +133,7 @@ public struct TextViewProperties {
         var end: Int
     }
     
-    enum TextAlignment: String, Decodable {
+    enum TextAlignment: String {
         case left
         case right
         case center
@@ -155,7 +155,7 @@ public struct TextViewProperties {
 }
 
 public struct FlexLayoutProperties {
-    enum ItemAlignment: String, Decodable {
+    enum ItemAlignment: String {
         case stretch
         case center
         case flexStart  = "flex-start"
@@ -163,7 +163,7 @@ public struct FlexLayoutProperties {
         case baseline
     }
     
-    enum ContentJustification: String, Decodable {
+    enum ContentJustification: String {
         case flexStart      = "flex-start"
         case flexEnd        = "flex-end"
         case center         = "center"
@@ -171,7 +171,7 @@ public struct FlexLayoutProperties {
         case spaceAround    = "space-around"
     }
     
-    enum Direction: String, Decodable {
+    enum Direction: String {
         case row
         case column
     }
@@ -181,14 +181,9 @@ public struct FlexLayoutProperties {
     var contentJustification: ContentJustification = .flexStart
 }
 
-public struct ImageViewProperties: Decodable {
+public struct ImageViewProperties {
     var source: URL
     var caption: String? = nil// TODO: apply to imageviews
-    
-    enum CodingKeys: String, CodingKey {
-        case source = "src"
-        case caption = "label"
-    }
 }
 
 public struct VideoViewProperties {
@@ -262,7 +257,7 @@ public struct Shadow {
 }
 
 public struct Stroke {
-    public enum Style: String, Decodable {
+    public enum Style: String {
         case solid
         case dotted
         case dashed
@@ -307,7 +302,7 @@ extension Transform where Value: Numeric {
     }
 }
 
-enum HorizontalGravity: String, RawRepresentable, Decodable {
+enum HorizontalGravity: String, RawRepresentable {
     case left   = "left_horizontal"
     case center = "center_horizontal"
     case right  = "right_horizontal"
@@ -319,20 +314,20 @@ public struct Accessibility {
 }
 
 public struct BackgroundImage {
-    public enum ScaleType: String, Decodable {
+    public enum ScaleType: String {
         case none // original size
         case centerCrop = "center_crop" // fill
         case centerInside = "center_inside" // fit
     }
     
-    public enum TileMode: String, Decodable {
+    public enum TileMode: String {
         case none
         case repeatX = "repeat_x"
         case repeatY = "repeat_y"
         case repeatXY = "repeat"
     }
     
-    public enum Position: String, Decodable {
+    public enum Position: String {
         case leftTop = "left_top"
         case leftCenter = "left_center"
         case leftBottom = "left_bottom"
